backend/requirements.txt
# fastapi
fastapi
uvicorn
python-multipart
python-dotenv
itsdangerous~=2.1.2
passlib==1.7.4
python-jose
pydantic~=2.7.0
pydantic_core~=2.18.1
pydantic-settings==2.3.4
bcrypt==4.0.1

# database
sqlalchemy
psycopg2-binary

# testing
pytest
httpx
